Add setting to control implementations code lens
